requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.0.1
idna==3.10
playwright==1.40.0
pyee==11.0.1
python-dotenv==1.0.0
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
